refactor(@angular/cli): modularize `find_examples` MCP tool
The `find_examples` MCP tool has been refactored to improve maintainability and organization.
The original `examples.ts` file was broken down into several smaller, single-responsibility
modules within `packages/angular/cli/src/commands/mcp/tools/examples/`.
